requirements-dev.txt
watchdog[watchmedo]==4.0.0
PyGithub==2.3.0
loguru==0.7.2
